reboot the BBB when the "MMC/sysfs" error is detected
Some devices do occasionally fail to boot. Whenever this happens,
we found a suspicious message like this in the journal:

"
/sys/devices/platform/ocp/481d8000.mmc/mmc_host/mmc1/mmc1:0001/block/mmcblk1
and
/sys/devices/platform/ocp/481d8000.mmc/mmc_host/mmc1/mmc1:0001/block/mmcblk1/mmcblk1boot0d8000.mmc.device
appeared twice with different sysfs paths ...
"

We do not know the reason for the issue nor are we able to solve it. So we decided to add a workaround:
Whenever the message is detected in the journal, reboot the BeagleBone. If the error
is found immediately after such a rescue-reboot, do not reboot again to avoid endless reboot-loops.

Additionally, we create a file /HANDLED_MMC_DEVICE_APPEARED_TWICE.log which will
have an entry for each error detected to analyse devices post-mortem.