Fight glitches on ePC
The ePC has 2 CPUs with 2 cores, each.

CPU 0 owns core 0 and core 2.
CPU 1 owns core 1 and core 3.

Our audio thread is bound to core 0.
The kernel schedules other process on all the cores. When a heavy
process is scheduled on core 0 or core 2, the audio thread may be
interrupted. So we steal those cores from scheduler by passing
an according kernel parameter. (isolcpus=0,2)

MidiThread so far has been bound to core 1. Now, we bind it to core 2
which won't do anything despite of handling midi, now.