When undo/redo-ing a lot of actions, each of the actions triggered HWUI::setupFocusAndMode, even though only the last of those setup calls wins the race. By introducing a DelayedJob, all in-between calls to setupFocusAndMode will be ignored now.
Also, each parameter selection created a new PanelUnitParameterEditMode,
just to throw it away in setUsageMode, if it is of the same typ.
Creating a UnitMode means also subscribing to signals, some of them delayed.
Teaming with the behaviour described above, undoing of multiple actions
finally created dozens of UsageMode, just to throw them away immediately.
This patch will optimize this misbehaviour.