Fix 'Undo restores wrong sound'
Example:
+-----------+-------------------+---------------------------------+
| Timepoint | User Interaction  |        AE param changes         |
+-----------+-------------------+---------------------------------+
| (I)       | Param A (0...1)   | Param A (0...1)                 |
| (II)      | Param A (1...0)   | Param A (1...0)                 |
| (III)     | UndoJump to (I)   | Param A (0...1),Param A (1...0) |
| (IV)      | UndoJump to (III) | Param A (0...1),Param A (1...0) |
+-----------+-------------------+---------------------------------+

On parameter changes, the AE used to compare the new target value with
the *current value* instead of the *old target value*.
If, as in (III) and (IV), two contradicting requests drop in quickly
enough, the AE did not trigger a new ramp to the new target, but keeps
ramping towards the old target value.
This patch fixes the problem by not comparing at all, but simply update
the always-running ramp parameters accordingly.