refactored ToggleTestTone functionality (C15Synth, DspInterface)
The mechanism should be usable from the C15 CLI ("t"), but the PlayController should also be able to use the feature. Therefore, a virtual method in the DspInterface has been added. The DspHost overrides it and applies the mechanism.

Similarly, a virtual onSettingInitialSinglePreset method has been provided, allowing developers to quickly load a simple sound with audible output.