issue-3047: Fix race condition between sshd on bbb and ssh on epc
Sometimes, in particular after an update, the sshd on beagle bone takes
some time to start (maybe because of dicing a new host key?).
Any attempts to connect to the server are denied during that phase.
When epc tries to ssh-mount the USB stick connected to BBB, it only
tries once. If this try failes because sshd is not yet ready,
Epc has no access to the USB stick.
This patch adds a dependency to playground service in order to poll
the ssh service at BBB for at most 30s and then brute-force mounts
the usb stick. If mounting fails (because previous attempt already
suceeded), the error is just ignored (note the '-' sign).