toggling the test tone should always produce logs
This was recently wrapped in a compile time (constexpr bool LOG_SETTINGS), logging only when enabled. These "flags" are intended to ease development, but should be disabled for production.

However, logging Test Tone Events is crucial and should not depend on that (it also is virtually unreachable for the user). So, Test Tone Events now again always produce logs (from the Command Line and from PlayController Events).