introduce epc3 os
some NUCs of 11i3 series do not shutdown/halt properly if no hdmi device
is connected. We could not investigate the exact reason for this,
but with an up-to-date os, the problem disappears, so we decided to
introduce a new os for all NUCs >= 11i3.
When introducing the new os, we also decided that a better separation
between os and NL software is desireable to avoid multiple compilation
and packaging of NL software components. Thus, this patch for now only
introduces the 'real' os, basically as 'epc3-base-meta' package.
The new package defines all dependent packages that 'define' the NonLinux3
distribution and does the required tweaks in the post-install script.
No bootfs/rootfs images are required anymore, as installation of a
NonLinux3 on a new NUC is simply done by installing the new package,
which will in turn install all the dependencies and tweak the system
to our needs afterwards.

Additionally, the folder build-tools/epc-one-for-all replaces all
createUpdate routines for epc1/2/3 with a single rootfs created by the
abovementioned package. Out future updates will contain only one rootfs
tar that replaces the whole installed os, except the kernel modules.