issue-3392: Downgrade wpa_supplicant
Mac os devices (>= Monterey) cannot connect to C15 running
wpa_supplicant >= 2.10. So we have to downgrade to 2.9 which
seems to work fine.
While working on this, I found that I have to pin arch updates
to a dedicated version (2022-04-24) as otherwise due to the
rolling distribution nature of arch linux each rebuilding
of the docker container will fetch the newest version of
everything.

Probably, this pinning will introduce the known 'signature'
errors in the future, which I will solve then. (It is hard to
fix/test it now, as currently the errors do not yet exist)