Fix poll return value handling
The implemented handling off poll result was just
wrong, as it compared the 'number of evented file handles
or errno' with 'poll event type'. This used to work, but
only accidentally.
Now, we throw on every possible error (EFAULT, EINTR, EINVAL, ENOMEM)
as there is nothing we could do to fix if any of them is returned.

OTOT, a small refactoring is applied to MidiBridge IO::fetchChunk
as the 'magic' "+1" should be used a local as possible.